A lot of updates this time
The Dungeon Randomizer is underway, I'm back in business. I have made all the room prefabs for the Dungeons and I am extending the development time to the end of the year.
My plan is to charge between 1-3 € for the game, hopefully the game will be good enough for the price tag.

Anyways I'll hopefully be continuing work now.