The Generation WORKS hahahahahha
Anyways the generation works great. I have moved on to enemy spawners but they don't work... Why am I working on this again? Don't know. Opened the project and was like fuck it. Let's make something of it. So the hiatus is now on the other game(Shinobi).

I'm back bitch