AddedBonusSpritesAndOtherSprites
Need to add spawn points for objects, create bonuses add nodes for pathfinding, add visual detection to enemies, finish the weapon platform for ease of use and add levels to the level generator, to streamline the experience. Basically a bunch of work but should be done in 2 months if I try real hard.